Move API calls and form elements to dashboard page
